Separated the tests for Factor, TabularCPD and JointProbabilityDistribution classes
